make the dialog reload when upload image
