+++ b2d-venture/app/dashboard/business/[id]/dataroom/page.tsx
@@ -9,8 +9,11 @@ export default function Page({ params }) {
   const {id} = params;
   const [files, setFiles] = useState([])
   const [user_id, setUserId] = useState("")
+  const [isLoading, setIsLoading] = useState(true)
+
   async function fetchData() {
     try {
+      setIsLoading(true);
       const response = await fetch(`/api/fetchingData/DataRoom/${id}`);
       const data = await response.json();
       
@@ -18,6 +21,7 @@ export default function Page({ params }) {
       const business = await response1.json();
       setFiles(data.data.files || []);
       setUserId(business.user_id || "");
+      setIsLoading(false)
     } catch (error) {
       console.error("Failed to fetch data:", error);
     }
@@ -27,18 +31,33 @@ export default function Page({ params }) {
     fetchData()
   }, [])
   
+ 
   return(
     // OnUploadComplete={}
     <div className="flex overflow-hidden">
         <div className="flex ml-10 w-[40vw] h-screen">
-        <DragAndDrop type="dataroom" className={"flex mt-44 bg-transparent justify-center h-screen w-screen"} onUploadComplete={fetchData}/>
+        <DragAndDrop 
+          type="dataroom" 
+          onUploadComplete={fetchData}
+          className={"flex mt-44 bg-transparent justify-center h-screen w-screen"} 
+          />
         </div>
 
         <div className="absolute right-0 flex w-[40vw] h-screen bg-white shadow-lg  ">
             <div className="flex mt-20 ">
                 <div className="ml-16 mt-5">
                     <h1 className="mb-3 text-2xl font-semibold">Data Room</h1>
-                      {files.map((file,index)=>(
+                      {isLoading ? (
+                          <div className="">
+                            <img
+                                src="/assets/icons/icons-loading.gif"
+                                alt="loading"
+                                className="object-contain"
+                              />
+                            <div className="loader">Loading...</div> 
+                          </div>
+                      )
+                      : files.map((file,index)=>(
                         <FileContainer 
                           key={index} 
                           name={file.name}
