+++ b2d-venture/components/shared/BusinessDashboard/DragAndDrop.tsx
@@ -8,7 +8,7 @@ import { useRouter } from 'next/router';
 import toast from "react-hot-toast";
 import { Toaster } from 'react-hot-toast';
 
-export default function DragAndDrop({type, className, OnUploadComplete}) {
+export default function DragAndDrop({type, className, onUploadComplete}) {
   const [dragActive, setDragActive] = useState<boolean>(false);
   const inputRef = useRef<any>(null);
   const [files, setFiles] = useState<any>([]);
@@ -53,10 +53,10 @@ export default function DragAndDrop({type, className, OnUploadComplete}) {
       if (!res.ok){
         throw new Error('Upload failed.');
       }
-
+      onUploadComplete();
       const result = await res.json();
       toast.success('File uploaded successfully!');
-      OnUploadComplete();
+      
       console.log('Upload successful:', result);
       // reset after successful
       setFiles([]);
